Added visualisation for model params, renamed model_parameters
